# pkg/docs/swagger.yaml
basePath: /dragonfly
definitions:
...
  rest.SimpleMsg:
    properties:
      message:
        example: Any message
        type: string
    type: object
...
  types.AlertTask:
    properties:
      alert_event_message:
        type: string
      alert_event_name:
        type: string
      alert_event_type:
        type: string
      alert_math_expression:
        type: string
      alert_threshold:
        type: number
      critic_event_cnt:
        type: integer
      event_duration:
        type: string
      measurement:
        type: string
      metric:
        type: string
      name:
        type: string
      target_id:
        type: string
      target_type:
        type: string
      warn_event_cnt:
        type: integer
    type: object
host: localhost:9090
info:
  contact:
    email: contact-to-cloud-barista@googlegroups.com
    name: API Support
    url: http://cloud-barista.github.io
  description: CB-Dragonfly REST API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: CB-Dragonfly REST API
  version: "1.0"
paths:
...
  /alert/task:
    post:
      consumes:
      - application/json
      description: 알람 생성
      parameters:
      - description: Details for an Event object
        in: body
        name: eventHandlerInfo
        required: true
        schema:
          $ref: '#/definitions/types.AlertTask'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.AlertTask'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
      summary: Create monitoring alert
      tags:
      - '[Alarm] Alarm management'
  /alert/task/{task_id}:
    delete:
      consumes:
      - application/json
      description: 알람 삭제
      parameters:
      - description: 태스크 아이디
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
      summary: Delete monitoring alert
      tags:
      - '[Alarm] Alarm management'
    get:
      consumes:
      - application/json
      description: 알람 조회
      parameters:
      - description: 태스크 아이디
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.AlertTask'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
      summary: Get monitoring alert
      tags:
      - '[Alarm] Alarm management'
    put:
      consumes:
      - application/json
      description: 알람 수정
      parameters:
      - description: 태스크 아이디
        in: path
        name: task_id
        required: true
        type: string
      - description: Details for an Event object
        in: body
        name: eventHandlerInfo
        required: true
        schema:
          $ref: '#/definitions/types.AlertTask'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.AlertTask'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
      summary: Update monitoring alert
      tags:
      - '[Alarm] Alarm management'
...
  /alert/tasks:
    get:
      consumes:
      - application/json
      description: 알람 목록 조회
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.AlertTask'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.SimpleMsg'
      summary: List monitoring alert
      tags:
      - '[Alarm] Alarm management'
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"